backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.1.0
gunicorn==21.2.0
